chore: move `color_preference` from `crate::types` to `crate::types::near` in `cargo-near-build`
