+++ cargo-near-build/src/env_keys.rs
@@ -16,6 +16,7 @@ pub mod nep330 {
     pub const LINK: &str = "NEP330_LINK";
     pub const VERSION: &str = "NEP330_VERSION";
     // ====================== End section =======================================
+    // TODO: make mod pub(crate)
     pub mod nonspec {
         pub const SERVER_DISABLE_INTERACTIVE: &str = "CARGO_NEAR_SERVER_BUILD_DISABLE_INTERACTIVE";
     }
