+++ cargo-near-build/src/types/near/color_preference.rs
