+++ cargo-near-build/src/types/near/mod.rs
@@ -1,5 +1,6 @@
 pub mod abi;
 pub mod build;
 pub mod build_extended;
+pub mod color_preference;
 #[cfg(feature = "docker")]
 pub mod docker_build;
